Added new submodule. Changed initial_window fh in splitters
