+++ sktime/forecasting/model_evaluation/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3 -u
+# -*- coding: utf-8 -*-
+# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
+
+__author__ = ["Martin Walter"]
+__all__ = ["evaluate"]
+
+from sktime.forecasting.model_evaluation._functions import evaluate
